[None][refactor] Move MmEncoderMixin to modeling_multimodal_encoder
PR review (2ez4bz): the mixin lives next to multimodal embedding
helpers in `modeling_multimodal_utils.py` -- it should sit with the
other encoder-shaped classes. Move it into the existing
`modeling_multimodal_encoder.py` (which already hosts VILA's
`VisionTower` wrappers) and update every import site:
- modeling_clip / pixtral / qwen2vl / qwen3vl / radio / siglip
- pyexecutor/model_engine
- tests/unittest/_torch/modeling/test_mm_encoder_mixin
