+++ tensorrt_llm/_torch/models/modeling_clip.py
@@ -16,7 +16,7 @@ from ..model_config import ModelConfig
 from ..modules.attention import Attention
 from ..modules.mlp import MLP
 from .hf_parameter_utils import get_parameter_device, get_parameter_dtype
-from .modeling_multimodal_utils import MmEncoderMixin
+from .modeling_multimodal_encoder import MmEncoderMixin
 from .modeling_utils import _load_weights_impl, register_auto_model
