+++ tensorrt_llm/_torch/models/modeling_multimodal_encoder.py
@@ -16,11 +16,35 @@
 # This file is based on official VILA: https://github.com/NVlabs/VILA/blob/main/llava/model/multimodal_encoder/
 
 import os
+from typing import Optional, Type
 
 import torch.nn as nn
 from transformers import AutoConfig, AutoModel
 
+from ..attention_backend.interface import AttentionMetadata
 from .modeling_multimodal_utils import multiscale_forward
+
+
+class MmEncoderMixin:
+    """Marker + default `setup_attn_metadata` for multimodal encoders whose
+    AttentionMetadata is built by `PyTorchModelEngine` after model load using
+    runtime sizes (`max_batch_size`, `max_num_tokens`).
+
+    Subclasses set `metadata_cls` in their own `__init__` (typically from
+    `get_attention_backend(model_config.attn_backend).Metadata`) and either
+    use the default `setup_attn_metadata` below or override it for custom
+    Metadata kwargs (e.g. FlashInfer `kv_layout`, multi-metadata encoders).
+    """
+    metadata_cls: Type[AttentionMetadata]
+    attn_metadata: Optional[AttentionMetadata] = None
+
+    def setup_attn_metadata(self, max_num_requests: int,
+                            max_num_tokens: int) -> None:
+        self.attn_metadata = self.metadata_cls(
+            max_num_requests=max_num_requests,
+            max_num_tokens=max_num_tokens,
+            kv_cache_manager=None,
+        )
 
 
 class VisionTower(nn.Module):
