+++ tensorrt_llm/_torch/models/modeling_multimodal_utils.py
@@ -19,7 +19,7 @@
 import functools
 import math
 import os
-from typing import Any, Callable, Dict, List, Optional, Tuple, Type, Union, cast
+from typing import Any, Callable, Dict, List, Optional, Tuple, Union, cast
 
 import torch
 import torch.nn.functional as F
@@ -27,34 +27,11 @@ from einops import rearrange
 from PIL import Image
 from torchvision.transforms import Normalize, Resize, ToTensor
 
-from tensorrt_llm._torch.attention_backend.interface import AttentionMetadata
 from tensorrt_llm._torch.modules.embedding import Embedding
 from tensorrt_llm.inputs.multimodal import MultimodalParams
 from tensorrt_llm.logger import logger
 
 _MULTIMODAL_ENV_NAME = "TLLM_MULTIMODAL_DISAGGREGATED"
-
-
-class MmEncoderMixin:
-    """Marker + default `setup_attn_metadata` for multimodal encoders whose
-    AttentionMetadata is built by `PyTorchModelEngine` after model load using
-    runtime sizes (`max_batch_size`, `max_num_tokens`).
-
-    Subclasses set `metadata_cls` in their own `__init__` (typically from
-    `get_attention_backend(model_config.attn_backend).Metadata`) and either
-    use the default `setup_attn_metadata` below or override it for custom
-    Metadata kwargs (e.g. FlashInfer `kv_layout`, multi-metadata encoders).
-    """
-    metadata_cls: Type[AttentionMetadata]
-    attn_metadata: Optional[AttentionMetadata] = None
-
-    def setup_attn_metadata(self, max_num_requests: int,
-                            max_num_tokens: int) -> None:
-        self.attn_metadata = self.metadata_cls(
-            max_num_requests=max_num_requests,
-            max_num_tokens=max_num_tokens,
-            kv_cache_manager=None,
-        )
 
 
 # Make this a runtime lookup rather than a module-wide constant for easier unit testing.
