+++ tensorrt_llm/_torch/models/modeling_siglip.py
@@ -13,7 +13,7 @@ from ..attention_backend.utils import get_attention_backend
 from ..model_config import ModelConfig
 from .hf_parameter_utils import get_parameter_device, get_parameter_dtype
 from .modeling_clip import CLIPEncoder
-from .modeling_multimodal_utils import MmEncoderMixin
+from .modeling_multimodal_encoder import MmEncoderMixin
 from .modeling_utils import _load_weights_impl, register_auto_model
 
 SiglipEncoder = CLIPEncoder
