+++ tensorrt_llm/_torch/pyexecutor/model_engine.py
@@ -49,8 +49,8 @@ from ..expert_statistic import ExpertStatistic
 from ..memory_buffer_utils import clear_memory_buffers, with_shared_pool
 from ..metadata import KVCacheParams
 from ..models.checkpoints.base_checkpoint_loader import BaseCheckpointLoader
-from ..models.modeling_multimodal_utils import (MmEncoderMixin,
-                                                filter_mm_token_from_input_ids)
+from ..models.modeling_multimodal_encoder import MmEncoderMixin
+from ..models.modeling_multimodal_utils import filter_mm_token_from_input_ids
 from ..models.modeling_utils import DecoderModelForCausalLM
 from ..modules.fused_moe.moe_load_balancer import (MoeLoadBalancer,
                                                    MoeLoadBalancerIterContext)
